+++ zotero-plugin/bootstrap.js
@@ -5,6 +5,7 @@ function install() {}
 async function startup({ id, version, rootURI }) {
   await Zotero.initializationPromise;
   Services.scriptloader.loadSubScript(rootURI + "native-exporter.js");
+  Services.scriptloader.loadSubScript(rootURI + "destinations.js");
   Services.scriptloader.loadSubScript(rootURI + "zpm.js");
   await ZPMPlugin.startup({ id, version, rootURI });
 }